requirements.txt
#base
alembic==1.10.2
SQLAlchemy==2.0.7
fastapi==0.95.0
pydantic==1.10.7
uvicorn==0.21.1
asyncpg==0.27.0
python-dotenv==1.0.0
python-jose==3.3.0
bcrypt==4.0.1
passlib==1.7.4

#tests
pytest==7.2.2
pytest-asyncio==0.21.0
httpx==0.23.3
pytest-env==0.8.1